+++ src/mol-repr/volume/dot.ts
@@ -292,7 +292,7 @@ function getDotLoci(pickingId: PickingId, volume: Volume, key: number, props: Vo
         const instances = OrderedSet.ofSingleton(instanceId as Volume.InstanceIndex);
         if (granularity === 'volume') {
             return Volume.Loci(volume, instances);
-        } else if (granularity === 'object') {
+        } else if (granularity === 'object' || groupId === PickingId.Null) {
             return Volume.Isosurface.Loci(volume, props.isoValue, instances);
         } else {
             const indices = Interval.ofSingleton(groupId as Volume.CellIndex);