+++ CHANGELOG.md
@@ -106,6 +106,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix `ColorScale` for continuous case without offsets (broke in v4.13.0)
 - Experimental: support for custom color themes in Sequence Panel
 - Switch files.rcsb.org validation report URL to new endpoint /validation/view
+- Improve picking of objects with too many groups, pick whole instance/object
 
 ## [v4.18.0] - 2025-06-08
 - MolViewSpec extension: